BitemporalPostgres.jl

BitemporalPostgres.Component — Type

Component

an abstract component of a bitemporal entity i.e. a contract, a reference to a partner a component's history of states is represented by its revisions

BitemporalPostgres.ComponentRevision — Type

ComponentRevision

a component of a bitemporal entity i.e. a contract, a reference to a partner a component's history of states is represented by its revisions

BitemporalPostgres.History — Type

History

The root of an aggregate entity, which groups versions and entities

BitemporalPostgres.Node — Type

Node

node of tree of mutations

BitemporalPostgres.Role — Type

Role

role of a relationship

BitemporalPostgres.SubComponent — Type

SubComponent

an abstract subcomponent of a bitemporal entity i.e. product item of a contract

BitemporalPostgres.TestDummyComponent — Type

TestDummyComponent <: Component

a component of a bitemporal entity

BitemporalPostgres.TestDummyComponentRevision — Type

TestDummyComponentRevision <: ComponentRevision

a revision of a contract component of a bitemporal entity

BitemporalPostgres.TestDummySubComponent — Type

TestDummySubComponent <: Component a component of a bitemporal entity

BitemporalPostgres.TestDummySubComponentRevision — Type

TestDummySubComponentRevision <: ComponentRevision

a revision of a contract component of a bitemporal entity

BitemporalPostgres.ValidityInterval — Type

ValidityInterval

a 2-dimensional interval of validity of a version with respect to the state

of the data base storage and

of the world

See the theory page for background

BitemporalPostgres.Version — Type

Version

version number of the state of aggregate entity

BitemporalPostgres.Workflow — Type

Workflow Holds application based transaction data.

current version

world validFrom date

db validfrom date

is_committed means: bitemporal transaction is

pending: 0

committed: 1

rolled back: 2

BitemporalPostgres.commit_workflow! — Method

commitworkflow!(w::Workflow) commits the bitemporal transaction identified by refversion begun with the creation of w

sets w's version's validityInterval to is_committed

ends all overlapping intervals to tsdbinvalidfrom w.tsdbvalidfrom

creates a new interval for the non_overlapping part of a partially interval if any

BitemporalPostgres.create_component! — Method

create_component!(c::Component, cr :: ComponentRevision, w:: Workflow)

creates a component and a componentRevision

BitemporalPostgres.create_entity! — Method

create_entity!(w::Workflow)

opens a bitemporal transaction identified by ref_version

persists a history, version, validityInterval and a Workflow

requires: w.tsw_validfrom is a valid date

BitemporalPostgres.create_subcomponent! — Method

create_subcomponent!(c::Component, sc::SubComponent, cr :: ComponentRevision, w:: Workflow) creates a subcomponent and a componentRevision

BitemporalPostgres.delete_component! — Method

delete_component!(c::T, w::Workflow) where {T<:Component}

deletes a component if it was created for the current version or

mark its latest component revision as invalid

BitemporalPostgres.findcomponentrevision — Method

findcomponentrevision(t::Type{T},refcomponent::DbId,refversion::DbId,)::Vector{T} where {T<:ComponentRevision} retrieves the version_id of a bitemporal history asof tsdb as per tsw

BitemporalPostgres.findversion — Method

findversion(refhistory::DbId, tsdb::ZonedDateTime, tsw::ZonedDateTime)::DbId retrieves the versionid of a bitemporal history asof tsdb as per tsw

BitemporalPostgres.get_typeof_revision — Method

gettypeofrevision(component::TestDummySubComponent) :: T where {T<:ComponentRevision}

BitemporalPostgres.get_typeof_revision — Method

gettypeofrevision(component::TestDummyComponent) :: T where {T<:ComponentRevision}

BitemporalPostgres.get_typeof_revision — Method

gettypeofrevision(component::T) :: R where {T<: Component, R <: ComponentRevision} returns the actual subtype of ComponentRevision that fits the actual type of component

BitemporalPostgres.mkforest — Method

mkforest(hid::DbId,tsdbinvalidfrom::ZonedDateTime,tsworldvalidfrom::ZonedDateTime,tsworld_invalidfrom::ZonedDateTime,)::Vector{Node}

builds a tree of vectors of version nodes where

* nodes in one vector denote conecutive mutations and
* child node vectors denote mutations which have been retrospectively corrected by their predecessor

see: Theory: Textual representation of mutation histories

BitemporalPostgres.rollback_workflow! — Method

rollbackworkflow!(w::Workflow) rolls back the bitemporal transaction identified by refversion begun with the creation of w

BitemporalPostgres.update_component! — Method

update_component!(cr :: ComponentRevision, crNew :: ComponentRevision, w:: Workflow)

terminates a componentRevision and persists its successor

BitemporalPostgres.update_entity! — Method

update_entity!(w::Workflow)

opens a bitemporal transaction identified by ref_version

persists a version, a validityInterval and a Workflow

requires:

w.tsw_validfrom is a valid date

w.ref_history is a valid history id

w.refversion is a valid version of w.refhistory