BitemporalPostgres.DDL.jl

BitemporalPostgres.DDL.create_validity_intervals_constraints — Method

createvalidityintervals_constraints() Install the GIST constraint that prevents creation of overlapping intervals, i.e. that at most one version is valid for a given 2 dimensional point in transaction and reference time

BitemporalPostgres.DDL.create_versions_trigger — Method

createversionstrigger() propagate skalars refvalidfrom and refinvalidfrom to range ref_valid skalars are needed enable foreign key constraints, which in POSTGRES cannot be declared onto bounds of intervals directly.

BitemporalPostgres.DDL.down — Method

down() callback from SearchLight.Migrations.down tearing down the Database Schema for bitemproal Transactions including test dummy tables

BitemporalPostgres.DDL.up — Method

up() callback from SearchLight.Migrations.up creating the Database Schema for bitemporal Transactions including test dummy tables